MovieStore/Tests/MovieStore.UnitTests/Application/CustomerOperations/Commands/DeleteCustomer/DeleteCustomerCommandTests.cs


using FluentAssertions;
using MovieStore.Application.CustomerOperations.Commands.DeleteCustomer;
using MovieStore.DbOperations;
using MovieStore.UnitTests.TestSetup;
using System;
using Xunit;

namespace MovieStore.UnitTests.Application.CustomerOperations.Commands.DeleteCustomer
{
    public class DeleteCustomerCommandTests : IClassFixture<CommonTestFixture>
    {
        private readonly MovieStoreDbContext _context;
        public DeleteCustomerCommandTests(CommonTestFixture testFixture)
        {
            _context = testFixture.Context;
        }

        [Theory]
        [InlineData(10)]
        [InlineData(11)]
        [InlineData(12)]
        public void WhenCustomerNotFound_InvalidOperationException_ShouldBeReturn(int customerId)
        {
            DeleteCustomerCommand command = new DeleteCustomerCommand(_context);
            command.CustomerId = customerId;
            FluentActions.Invoking(() => command.Handle()).Should().Throw<InvalidOperationException>().And.Message.Should().Be("Silinecek müşteri bulunamadı");
        }
    }
}
